requirements.txt
astroid==2.15.0
async-timeout==4.0.2
attrs==22.2.0
bandit==1.7.3
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
coverage==7.1.0
dill==0.3.6
distlib==0.3.6
exceptiongroup==1.1.1
filelock==3.11.0
flake8==5.0.4
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
idna==3.4
importlib-metadata==6.2.0
iniconfig==2.0.0
isort==5.11.5
joblib==1.2.0
lazy-object-proxy==1.9.0
llvmlite==0.39.1
mando==0.6.4
mccabe==0.7.0
mypy==0.990
mypy-extensions==1.0.0
numba==0.56.4
numpy==1.21.0
packaging==23.0
pandas==1.2.5
pathspec==0.11.0
pbr==5.11.1
platformdirs==3.1.1
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.9.1
pyenv-win==3.1
pyflakes==2.5.0
pylint==2.17.0
pytest==7.2.0
pytest-cov==3.0.0
pytest-mock==3.9.0
python-dateutil==2.8.2
python-dotenv==0.18.0
pytz==2022.7.1
PyYAML==5.4.1
radon==5.1.0
redis==4.5.4
requests==2.28.2
scikit-learn==1.2.0
scipy==1.10.1
shap==0.41.0
six==1.16.0
slicer==0.0.7
smmap==5.0.0
spotipy==2.23.0
SQLAlchemy==1.4.31
stevedore==5.0.0
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
tox==3.24.3
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
virtualenv==20.21.0
wrapt==1.15.0
xenon==0.8.0
zipp==3.15.0
